requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1
django-ckeditor==6.7.1
django-import-export==4.1.1
django-js-asset==2.2.0
pillow==10.4.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
